feat(TCOMP-2100): refactor and cleanup
