--- component-server-parent/component-server-model/src/main/java/org/talend/sdk/component/server/front/model/icons/IconSymbol.java
@@ -1,43 +1,0 @@
-/**
- * Copyright (C) 2006-2024 Talend Inc. - www.talend.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.talend.sdk.component.server.front.model.icons;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class IconSymbol {
-
-    String icon;
-
-    String family;
-
-    String type;
-
-    String connector;
-
-    String theme;
-
-    byte[] content;
-
-    public String getUid() {
-        final String fmt = connector.isEmpty() ? "%s-%s-%s-%s" : "%s-%s-%s-%s-%s";
-        return String.format(fmt, theme, type, family, icon, connector);
-    }
-}
